se mueve!!!! pero solo en un cliente:(

+++ src_files/src/src_server/server_game/server_mapa/entidades/edificios/edificios_mapa.h
@@ -8,7 +8,7 @@ class Edificio : public Entidades {
 protected:
     int dimension_x;
     int dimension_y;
-    uint16_t id_jugador;
+    uint8_t id_jugador;
     uint8_t tipo_entidad;
 
 
@@ -17,7 +17,7 @@ public:
     virtual int obtenerDimensionX();
     virtual int obtenerDimensionY();
     virtual char obtenerIdentificador();
-    virtual uint16_t obtenerIdJugador();
+    virtual uint8_t obtenerIdJugador();
     virtual char obtenerTipoDeEntidad();
     virtual ~Edificio() = default;
 
@@ -29,49 +29,49 @@ public:
 
 class Silo: public Edificio {
 public:
-    Silo(YAML::Node& edificio_config, uint16_t id_jugador);
+    Silo(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~Silo() = default;
 };
 
 class FabricaLigera: public Edificio {
 public:
-    FabricaLigera(YAML::Node& edificio_config, uint16_t id_jugador);
+    FabricaLigera(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~FabricaLigera() = default;
 };
 
 class FabricaPesada: public Edificio {
 public:
-    FabricaPesada(YAML::Node& edificio_config, uint16_t id_jugador);
+    FabricaPesada(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~FabricaPesada() = default;
 };
 
 class Palacio: public Edificio {
 public:
-    Palacio(YAML::Node& edificio_config, uint16_t id_jugador);
+    Palacio(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~Palacio() = default;
 };
 
 class Refineria: public Edificio {
 public:
-    Refineria(YAML::Node& edificio_config, uint16_t id_jugador);
+    Refineria(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~Refineria() = default;
 };
 
 class Cuartel: public Edificio {
 public:
-    Cuartel(YAML::Node& edificio_config, uint16_t id_jugador);
+    Cuartel(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~Cuartel() = default;
 };
 
 class TrampaDeAire: public Edificio {
 public:
-    TrampaDeAire(YAML::Node& edificio_config, uint16_t id_jugador);
+    TrampaDeAire(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~TrampaDeAire() = default;
 };
 
 class CentroDeConstruccion: public Edificio {
 public:
-    CentroDeConstruccion(YAML::Node& edificio_config, uint16_t id_jugador);
+    CentroDeConstruccion(YAML::Node& edificio_config, uint8_t id_jugador);
     virtual ~CentroDeConstruccion() = default;
 };
