+++ src_files/src/src_server/server_game/server_mapa/entidades/unidades/unidades_mapa.h
@@ -8,7 +8,7 @@
 class UnidadesMapa : public Entidades {
 private:
     char parada_sobre;
-    uint16_t id_jugador;
+    uint8_t id_jugador;
     uint8_t id_unidad;
     uint8_t tipo_unidad;
     char tipo_entidad;
@@ -18,7 +18,7 @@ public:
     virtual char obtenerIdentificador();
     virtual char obtenerTerrenoQueEstaParada();
     virtual char obtenerTipoDeEntidad();
-    virtual uint16_t obtenerIdJugador();
+    virtual uint8_t obtenerIdJugador();
     virtual uint8_t obtenerTipoDeUnidad();
     virtual uint8_t obtenerIdUnidad();
     virtual ~UnidadesMapa() = default;
